feat: add new state, ownedState, and computedState APIs
