+++ packages/web-components/fast-element/package.json
@@ -50,9 +50,9 @@
       "types": "./dist/dts/utilities.d.ts",
       "default": "./dist/esm/utilities.js"
     },
-    "./hooks": {
-      "types": "./dist/dts/hooks.d.ts",
-      "default": "./dist/esm/hooks.js"
+    "./state": {
+      "types": "./dist/dts/state/exports.d.ts",
+      "default": "./dist/esm/state/exports.js"
     },
     "./context": {
       "types": "./dist/dts/context.d.ts",
