+++ packages/web-components/fast-element/src/state/exports.ts
@@ -1,0 +1,3 @@
+export { reactive } from "./reactive.js";
+export { watch } from "./watch.js";
+export * from "./state.js";
